+++ tests/test_ipython_formatter.py
@@ -8,8 +8,7 @@ test_yuuno
 Tests for `yuuno` module.
 """
 
-
-import sys
+import base64
 import unittest
 
 from IPython.testing import globalipapp
@@ -24,6 +23,11 @@ from yuuno.ipython.environment import load_ipython_extension, unload_ipython_ext
 from yuuno.ipython.formatter import Formatter, InlineFormat
 
 from test_png_output import SinglePixelFrame, TestPNGOutput
+
+
+class StaticObject(object):
+    pass
+STATIC_OBJECT = StaticObject()
 
 
 class TestClip(Clip):
@@ -41,7 +45,7 @@ class TestClipExtension(Extension):
         return True
 
     def initialize(self):
-        self.parent.registry.register(TestClip, int)
+        self.parent.registry.register(TestClip, StaticObject)
 
 
 class TestFormatter(unittest.TestCase):
@@ -55,6 +59,7 @@ class TestFormatter(unittest.TestCase):
 
         self.shell = globalipapp.get_ipython()
         load_ipython_extension(self.shell)
+        Yuuno.instance().output.icc_profile = None
 
         self.loaded = True
 
@@ -66,25 +71,21 @@ class TestFormatter(unittest.TestCase):
         tc = TestClip(None)
 
         inline = InlineFormat(environment=Yuuno.instance().environment, clip=tc)
-        self.assertEqual(inline.ipy_image().data, TestPNGOutput.EXPECTED_RESULT_SIMPLE)
+        self.assertEqual(inline.ipy_image.data, TestPNGOutput.EXPECTED_RESULT_SIMPLE)
 
     def test_002_formatter_lookup_success(self):
-        self.assertIsNotNone(self.shell.display_formatter.formatters['image/png'].lookup(1))
-        self.assertIsNotNone(self.shell.display_formatter.formatters['text/html'].lookup(1))
-        self.assertIsNotNone(self.shell.display_formatter.formatters['text/plain'].lookup(1))
+        self.assertIsNotNone(self.shell.display_formatter.formatters['image/png'].lookup(STATIC_OBJECT))
+        self.assertIsNotNone(self.shell.display_formatter.formatters['text/plain'].lookup(STATIC_OBJECT))
 
-        raw, _ = self.shell.display_formatter.formatters['image/png'].lookup(1)(1)
-        self.assertEqual(raw, TestPNGOutput.EXPECTED_RESULT_SIMPLE)
+        raw, _ = self.shell.display_formatter.formatters['image/png'].lookup(STATIC_OBJECT)(STATIC_OBJECT)
+        self.assertEqual(raw, base64.b64encode(TestPNGOutput.EXPECTED_RESULT_SIMPLE).decode('ascii') + '\n')
 
     def test_003_formatter_unload_success(self):
         self.loaded = False
         unload_ipython_extension(self.shell)
 
         with self.assertRaises(KeyError):
-            self.shell.display_formatter.formatters['image/png'].lookup(1)
+            self.shell.display_formatter.formatters['image/png'].lookup(STATIC_OBJECT)
 
         with self.assertRaises(KeyError):
-            self.shell.display_formatter.formatters['text/html'].lookup(1)
-
-        with self.assertRaises(KeyError):
-            self.shell.display_formatter.formatters['text/plain'].lookup(1)
+            self.shell.display_formatter.formatters['text/plain'].lookup(STATIC_OBJECT)