+++ yuuno/output/pil2png.py
@@ -16,7 +16,6 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-
 from io import BytesIO
 
 from traitlets import Unicode, CInt, Any
@@ -24,16 +23,20 @@ from traitlets.config import Configurable
 from PIL.Image import Image
 
 from yuuno.clip import Frame
+from yuuno.output.srgb_png import srgb
 
 
 class YuunoImageOutput(Configurable):
     """
     Defines an output for PNG-files
     """
+
+    ################
+    # Settings
     yuuno = Any(help="Reference to the current Yuuno instance.")
 
     zlib_compression: int = CInt(6, help="0=No compression\n1=Fastest\n9=Slowest", config=True)
-    icc_profile: str = Unicode(None, help="Specify the path to an ICC-Profile (Defaults to sRGB).", allow_none=True, config=True)
+    icc_profile: str = Unicode("sRGB", help="Specify the path to an ICC-Profile (Defaults to sRGB).", allow_none=True, config=True)
 
     def bytes_of(self, im: Frame) -> bytes:
         """
@@ -53,8 +56,11 @@ class YuunoImageOutput(Configurable):
             "format": "png"
         }
         if self.icc_profile is not None:
-            with open(self.icc_profile, "rb") as f:
-                settings["icc_profile"] = f.read()
+            if self.icc_profile != "sRGB":
+                with open(self.icc_profile, "rb") as f:
+                    settings["icc_profile"] = f.read()
+            else:
+                settings.update(srgb())
 
         f = BytesIO()
         im.save(f, **settings)