+++ yuuno/vs/clip.py
@@ -29,7 +29,8 @@ from vapoursynth import VideoNode, VideoFrame
 from yuuno import Yuuno
 from yuuno.clip import Clip, Frame
 from yuuno.vs.extension import VapourSynth
-from yuuno.vs.utils import get_proxy_or_core
+from yuuno.vs.utils import get_proxy_or_core, is_version
+from yuuno.vs.alpha import AlphaOutputClip
 
 
 def calculate_size(frame: VideoFrame, planeno: int) -> Tuple[int, int]:
@@ -104,7 +105,7 @@ def extract_plane(frame: VideoFrame, planeno: int, *, compat: bool=False) -> Ima
     :param compat:  Are we dealing with a compat format.
     :return: The extracted image.
     """
-    if get_proxy_or_core().version_number() > 36:
+    if is_version(36):
         return extract_plane_new(frame, planeno, compat=compat)
     else:
         return extract_plane_r36compat(frame, planeno, compat=compat)
@@ -112,11 +113,16 @@ def extract_plane(frame: VideoFrame, planeno: int, *, compat: bool=False) -> Ima
 
 class VapourSynthFrameWrapper(HasTraits, Frame):
 
+    pil_cache: Image.Image = Instance(Image.Image, allow_none=True)
+
     frame: VideoFrame = Instance(VideoFrame)
     compat_frame: VideoFrame = Instance(VideoFrame)
 
     def to_pil(self) -> Image.Image:
-        return extract_plane(self.compat_frame, 0, compat=True)
+        if self.pil_cache is None:
+            self.pil_cache = extract_plane(self.compat_frame, 0, compat=True)
+        # noinspection PyTypeChecker
+        return self.pil_cache
 
 
 class VapourSynthClipMixin(HasTraits):
@@ -181,3 +187,45 @@ class VapourSynthFrame(VapourSynthClipMixin, HasTraits):
     @observe("frame")
     def _frame_observe(self, value):
         self.clip = self._wrap_frame(value['new'])
+
+
+class VapourSynthAlphaFrameWrapper(HasTraits):
+    clip: VapourSynthFrameWrapper = Instance(VapourSynthFrameWrapper)
+    alpha: VapourSynthFrameWrapper = Instance(VapourSynthFrameWrapper)
+
+    _cache: Image.Image = Instance(Image.Image, allow_none=True)
+
+    @property
+    def color(self):
+        return clip[0]
+
+    def to_pil(self):
+        if self._cache is None:
+            color = self.clip.to_pil()
+            alpha = extract_plane(self.alpha.frame, 0)
+            color.putalpha(alpha)
+            self._cache = color
+        return self._cache
+
+
+class VapourSynthAlphaClip:
+
+    def __init__(self, clip):
+        if not isinstance(clip, AlphaOutputClip):
+            raise ValueError("Passed non Alpha-Clip into the wrapper")
+        
+        self.clip = VapourSynthClip(clip[0])
+        if clip[1] is None:
+            self.alpha = None
+        else:
+            self.alpha = VapourSynthClip(clip[1])
+
+    def __len__(self):
+        if self.alpha is None:
+            return len(self.clip)
+        return min(map(len, (self.clip, self.alpha)))
+
+    def __getitem__(self, item):
+        if self.alpha is None:
+            return self.clip[item]
+        return VapourSynthAlphaFrameWrapper(clip=self.clip[item], alpha=self.alpha[item])